requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-ninja==0.21.0
edupage-api==0.10.4
idna==3.4
mypy-extensions==1.0.0
orjson==3.8.10
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.7
requests==2.28.2
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
